cpu/esp32: initial support (#24)
* stm32: add initial uart support

* stm32: continued work on uart support

* UART improvements

* Improve uart support

* Update to newer version of GCC

* docker/test: use GCC 12.0.0

* makefile: run unit tests in docker

* github: do not run unit tests in container

* Improve UART implementation

* A bit of work on things

* docker/xtensa-esp32-elf: initial commit

* esp32: initial esp32 support

* Remove conditional block around alloca

* First initial functional blinky on ESP32

* Rename fried-badge to fri3d-badge

* Improve CI script

* archetypes/esp-idf: automatically download needed submodule

* ci: run unit tests for examples

* cpu/esp32: implement mock functions for esp-idf

* docker: update xtensa-esp32-elf docker image

* Fix building of esp projects under docker

* Fix execution of unit tests

* Attempt to fix ESP32 build

* Ignore licenses in dist/

* Add missing license files

* esp-idf docker container needs network

+++ source/mcud/util/format.d
@@ -25,7 +25,7 @@ private FormatSpec[] parseFormatString(string fmt)
 {
 	FormatSpec[] specs;
 	bool escape = false;
-	int argIndex = 0;
+	size_t argIndex = 0;
 	size_t escapeStart = 0;
 	FormatSpec spec;
 	for (size_t i = 0; i < fmt.length; i++)
